Add tests for DataListActions and restructure DataListItemActions to have more specific tests for its behavior
