+++ packages/components/src/DataList/components/DataListItemActions/DataListItemActions.test.tsx
@@ -67,59 +67,6 @@ describe("DataListItemActions", () => {
     expect(handleEditClick).toHaveBeenCalledTimes(1);
     expect(handleEditClick).toHaveBeenCalledWith({ id: 1 });
   });
-
-  it("should not render the more actions button if there are no more actions", () => {
-    mockItemActionComponent.mockReturnValueOnce(
-      <DataListItemActions>
-        <DataListAction icon="edit" label="Edit" />
-        <DataListAction icon="email" label="Email" />
-      </DataListItemActions>,
-    );
-    renderComponent();
-
-    expect(screen.queryByLabelText("More actions")).not.toBeInTheDocument();
-  });
-
-  it("should not render a 2nd action if the 2nd action doesn't have an icon even though the 3rd action have an icon", () => {
-    mockItemActionComponent.mockReturnValueOnce(
-      <DataListItemActions>
-        <DataListAction icon="edit" label="Edit" />
-        <DataListAction label="Email" />
-        <DataListAction icon="note" label="Note" />
-        <DataListAction label="Delete" />
-      </DataListItemActions>,
-    );
-    renderComponent();
-
-    expect(screen.getByLabelText("Edit")).toBeInTheDocument();
-    expect(screen.queryByLabelText("Email")).not.toBeInTheDocument();
-    expect(screen.queryByLabelText("Note")).not.toBeInTheDocument();
-    expect(screen.getByLabelText("More actions")).toBeInTheDocument();
-  });
-
-  it("should not render the first 2 actions if the first action doesn't have an icon", () => {
-    mockItemActionComponent.mockReturnValueOnce(
-      <DataListItemActions>
-        <DataListAction label="Edit" />
-        <DataListAction icon="email" label="Email" />
-        <DataListAction icon="note" label="Note" />
-        <DataListAction label="Delete" />
-      </DataListItemActions>,
-    );
-    renderComponent();
-
-    expect(screen.queryByLabelText("Edit")).not.toBeInTheDocument();
-    expect(screen.queryByLabelText("Email")).not.toBeInTheDocument();
-    expect(screen.queryByLabelText("Note")).not.toBeInTheDocument();
-    expect(screen.getByLabelText("More actions")).toBeInTheDocument();
-
-    const moreMenuButton = screen.getByLabelText("More actions");
-    userEvent.click(moreMenuButton);
-
-    const menu = screen.getByRole("menu");
-    expect(menu).toBeInTheDocument();
-    expect(within(menu).getAllByRole("button")).toHaveLength(4);
-  });
 });
 
 function renderComponent() {
